support double in custom and psycho skill effects

+++ src/main/java/pl/galajus/brokenpediabackend/user/skill/model/SkillCustomEffect.java
@@ -14,6 +14,6 @@ public class SkillCustomEffect {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String description;
-    private Integer value;
+    private Float value;
     private Long skillBasicId;
 }
